+++ src/Lib/Session.php
@@ -264,9 +264,16 @@ class Session
         return $out;
     }
 
-    public function getMyAmbassadorRegionIds(): array
+    public function getMyAmbassadorRegionIds(bool $includeWorkingGroups = true): array
     {
         $managedRegions = $this->getManagedRegions();
+
+        if (!$includeWorkingGroups) {
+            $managedRegions = array_filter($managedRegions, function ($region) {
+                return !UnitType::isGroup($region['type']);
+            });
+        }
+
         $out = [];
         foreach ($managedRegions as $region) {
             $out[] = $region['bezirk_id'];