+++ src/Modules/Region/RegionGateway.php
@@ -152,17 +152,6 @@ class RegionGateway extends BaseGateway
         );
     }
 
-    public function getFsAmbassadorIds(?int $foodsaverId): array
-    {
-        if ($foodsaverId === null) {
-            return [];
-        }
-
-        return $this->db->fetchAllValuesByCriteria('fs_botschafter', 'bezirk_id',
-            ['foodsaver_id' => $foodsaverId]
-        );
-    }
-
     public function listIdsForDescendantsAndSelf(int $regionId, bool $includeSelf = true, bool $includeWorkgroups = true): array
     {
         if ($regionId == RegionIDs::ROOT) {