+++ tests/unit/SearchGatewayTest.php
@@ -40,31 +40,31 @@ class SearchGatewayTest extends \Codeception\Test\Unit
 
         $this->assertEqualsCanonicalizing(
             [$f1, $f2],
-            array_column($this->searchGateway->searchUserInGroups('Albe', false, null), 'id')
+            array_column($this->searchGateway->searchUserInGroups('Albe', [], null), 'id')
         );
         $this->assertEqualsCanonicalizing(
             [$f4],
-            array_column($this->searchGateway->searchUserInGroups('Karl-Heinz', false, null), 'id')
+            array_column($this->searchGateway->searchUserInGroups('Karl-Heinz', [], null), 'id')
         );
         $this->assertEqualsCanonicalizing(
             [$f5],
-            array_column($this->searchGateway->searchUserInGroups('-(Matze)', false, null), 'id')
+            array_column($this->searchGateway->searchUserInGroups('-(Matze)', [], null), 'id')
         );
         $this->assertEqualsCanonicalizing(
             [$f5],
-            array_column($this->searchGateway->searchUserInGroups('von Heuschreckenland', false, null), 'id')
+            array_column($this->searchGateway->searchUserInGroups('von Heuschreckenland', [], null), 'id')
         );
         $this->assertEqualsCanonicalizing(
             [$f5],
-            array_column($this->searchGateway->searchUserInGroups('um Heuschreckenland', false, null), 'id')
+            array_column($this->searchGateway->searchUserInGroups('um Heuschreckenland', [], null), 'id')
         );
         $this->assertEqualsCanonicalizing(
             [],
-            array_column($this->searchGateway->searchUserInGroups('Fr*d', false, null), 'id')
+            array_column($this->searchGateway->searchUserInGroups('Fr*d', [], null), 'id')
         );
         $this->assertEqualsCanonicalizing(
             [$f2],
-            array_column($this->searchGateway->searchUserInGroups('Alb', false, [$region2['id']]), 'id')
+            array_column($this->searchGateway->searchUserInGroups('Alb', [], [$region2['id']]), 'id')
         );
     }
 }